+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #11 Upgrade to Bootique 0.13
 * #12 Support for init parameters for servlets and filters
+* #15 Support for sessions in Jetty
 
 ## 0.12: